requirements.txt
asgiref==3.9.1
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.10.0
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
psycopg==3.2.9
requests==2.32.4
sqlparse==0.5.3
